Fix/#5 : SellerProfile latitude/longitude를 BigDecimal로 변경
DB 컬럼은 DECIMAL(10,7)(고정소수점)인데 엔티티 필드는 Double(부동소수점)이면서
@Column(precision=10, scale=7)을 같이 쓰고 있었다. Hibernate 6.5는 부동소수점 타입에
precision/scale을 쓰면 "scale has no meaning for SQL floating point types"로 엔티티
메타데이터 빌드 자체를 실패시킨다. Render DB 스키마를 초기화하고 Flyway가 처음으로
끝까지 돈 뒤에야 이 에러가 드러났다.

API 스펙(DTO)은 그대로 Double을 유지하고, 서비스 레이어에서 BigDecimal로 변환한다.

+++ src/main/java/com/jejulocaltime/api/domain/seller/SellerProfile.java
@@ -7,6 +7,7 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -40,11 +41,14 @@ public class SellerProfile {
     // 사장님이 직접 등록/수정할 가게 및 정산 정보
     private String address;
     
+    // precision/scale은 DECIMAL(고정소수점) 컬럼에만 의미가 있어서 Double(부동소수점) 대신
+    // BigDecimal로 매핑해야 한다. Double + precision/scale 조합은 Hibernate 6.5에서
+    // "scale has no meaning for SQL floating point types"로 부팅 자체가 실패한다.
     @Column(precision = 10, scale = 7)
-    private Double latitude;
-    
+    private BigDecimal latitude;
+
     @Column(precision = 10, scale = 7)
-    private Double longitude;
+    private BigDecimal longitude;
 
     private String bankName;
     private String accountNumber;
